Setting default values on request failure
When a client request for a player state value fails, the client sets that value to a defualt; failure seems to occur when player is inactive, etc, when we want to clear those values/show default.
Added default values for PlayerButtonsState message in proto file.
processResponse is now testing for a response in an Envelope with an if statement instead of using assert; the Envelope comes from an external source, and although it should always have a response submessage when its type is set to response, this can't be guaranteed.
Added a request for track info on stop notification, for case where a playlist has just stopped playing.  The remote player will clear the track info display when done playing a playlist, and requesting the track on stop will cause this to happen for the client app.