Added player quitting notification to protocol
When the music player quits, it sends a message to the plug-in to notify it of the shutdown.  The plug-in was sending a stop notification to the clients, which respond with requests for track info, player position, and button state.  These requests cause the plug-in to restart the player, if the start player when messages are received option is enabled.  This made it impossible to stop the player while the plug-in was active.  To handle this a new notification message type to specify that the player is quitting was added.
Player sends quitting notification message when receving about to quit or quitting messages.
Client enters disconnected state when receiving player quitting notification.